pageable-tables/requirements.txt
flask
flask_sqlalchemy
flask_migrate
python-lorem
flake8
